requirements.txt
pymupdf==1.24.9         # fast PDF text extraction
tqdm==4.66.4            # progress bar
python-dotenv==1.0.1    # load OPENAI_API_KEY from .env
